requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
numpy==1.22.0
pandas==1.3.5
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.0
six==1.16.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
